Editing SMSBE - Customer (Sale)
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C53351

+++ TOneV2/Code/TOneV2/TOne.WhS.SMSBusinessEntity.Entities/CustomerSMSRateDraft/UpdateCustomerSMSDraftStatusInput.cs
@@ -8,7 +8,7 @@ namespace TOne.WhS.SMSBusinessEntity.Entities
 {
     public class UpdateCustomerSMSDraftStatusInput
     {
-        public int ProcessDraftID { get; set; }
+        public long ProcessDraftID { get; set; }
         public ProcessStatus NewStatus { get; set; }
     }
 }
